Animations, sound and attempted river death
Attach and adjusted the animations for the player and the obstacles (minus the turtle), Day2 attempted to implement the river death, tried raycasting this time, failed horribly (forgot to write this attempt for yesterday). But successfully added the sounds for each action and the background music when game is playing. Try to solve the river problem by tomorrow